Read me is added and execution option is provided with or without optimizer
